backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database - MongoDB
pymongo==4.10.1

# LangChain & LangGraph
langchain==0.3.15
langgraph==0.2.60

# AI/ML
emergentintegrations

# Document Processing
PyMuPDF==1.24.14
python-docx==1.1.2
openpyxl==3.1.5
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3

# Testing
pytest==8.3.4
httpx==0.28.1
